rename databricks test plugin to dbr350
