+++ test/features/org.talend.bigdata.se.test.feature/feature.xml
@@ -237,7 +237,7 @@
          fragment="true"
          unpack="false"/>
    <plugin
-         id="org.talend.hadoop.distribution.databricks.test"
+         id="org.talend.hadoop.distribution.dbr350.test"
          download-size="0"
          install-size="0"
          version="0.0.0"
