+++ test/plugins/org.talend.hadoop.distribution.databricks.test/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>org.talend.hadoop.distribution.databricks.test</name>
+	<name>org.talend.hadoop.distribution.dbr350.test</name>
 	<comment></comment>
 	<projects>
 	</projects>
